Move encrypt/decrypt to user security model
